pyproject.toml
[tool.poetry]
name = "ai_clinic"
version = "0.1.0"
description = ""
authors = ["Antony2500 <toxa241ff@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.111.0"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
cryptography = "^42.0.8"
pytest-async = "^0.1.1"
pytest = "^8.2.2"
pytest-postgresql = "^6.0.0"
pytest-asyncio = "^0.23.7"
bcrypt = "^4.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
testpaths = "tests"
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test* *Tests"
addopts = "-vv -x"
asyncio_mode = "auto"
pythonpath = ["."]
